Better download cache handling

+++ src/lib/services/file_downloader.py
@@ -46,10 +46,15 @@ class FileDownloader:
         else:
             self.__download_file(urls, save_dir, unzip=unzip, verbose=verbose)
 
-    def __download_file(self, url, save_dir, unzip=True, verbose=False):
+    def __download_file(self, url, save_dir, unzip=True, verbose=False, force=False):
         self.logger.info("Downloading file from url: {}".format(url))
         filename = os.path.basename(url)
         file_path = os.path.join(save_dir, filename)
+
+        if os.path.isfile(file_path) and not force:
+            self.logger.info("File {} already exists. Skipping".format(filename))
+            return
+
         with requests.get(url, stream=True) as r, open(file_path, "wb") as f:
             block_size = 1024
             total_size = int(r.headers.get('content-length', 0))
@@ -65,7 +70,7 @@ class FileDownloader:
 
         if can_be_decompressed(filename) and unzip:
             self.logger.info("Decompressing {}".format(file_path))
-            decompress(file_path)
+            decompress(file_path, delete_compressed=False)
 
 
     def __download_files(self, urls, save_dir, unzip=True, verbose=False):
